requirements.txt
aenum>=3.1.0
immutables>=0.16
ipinfo>=4.2.1
matplotlib>=3.4.2
numpy==1.21.1
pandas>=1.3.1
pillow>=8.3.1
pycryptodome==3.10.1
pygame>=2.0.1
PyInputPlus>=0.2.12
pyperclip>=1.8.2
PyQt5
rich>=10.7.0
screeninfo==0.7
traceback-with-variables==2.0.2
